feat: Add AbstractEntity to remove duplicate code
- 중복 코드를 줄이기 위해, AbstractEntity를 추가하였습니다.
- 현재 Question에만 적용하였습니다.
