+++ src/main/resources/data.sql
@@ -14,4 +14,4 @@ INSERT INTO USER (ID, USER_ID, PASSWORD, NAME, EMAIL) VALUES (13, 'freddy', '123
 INSERT INTO USER (ID, USER_ID, PASSWORD, NAME, EMAIL) VALUES (14, 'jane', '123', 'Jane', 'jane@jane.com');
 
 
-INSERT INTO QUESTION (ID, CONTENTS, DELETED, COUNT_OF_ANSWERS, CREATED_DATE_TIME, TITLE, WRITER_ID) VALUES (1, '삭제된 테스트 페이지', FALSE, 0, now(), '안녕하세요', 3)
+INSERT INTO QUESTION (ID, CONTENTS, DELETED, COUNT_OF_ANSWERS, CREATED_DATE_TIME, MODIFIED_DATE_TIME, TITLE, WRITER_ID) VALUES (1, '삭제된 테스트 페이지', FALSE, 0, now(), now(), '안녕하세요', 3)
